+++ example/bootstrap.php
@@ -8,17 +8,17 @@ $options = [
         'port' => getenv('QUEUE_PORT') ? : 11300,
     ],
     'tubes' => [
-        'Example1' => ['worker_num' => 2, 'class' => 'Codeages\\Plumber\\Example\\Example1Worker'],
-        'Example2' => ['worker_num' => 2, 'class' => 'Codeages\\Plumber\\Example\\Example2Worker'],
-        // 'ExampleForward' => [
-        //     'worker_num' => 2,
-        //     'class' => 'Codeages\Plumber\ForwardWorker',
-        //     'destination' => [
-        //         'host' => 'localhost', // 转发目的(host, port)队列，不能跟当前worker所监听的队列一样。
-        //         'port' => 11301,
-        //         'tubeName' => 'ExampleForward',
-        //     ]
-        // ],
+        'Example1' => ['worker_num' => 1, 'class' => 'Codeages\\Plumber\\Example\\Example1Worker'],
+        'Example2' => ['worker_num' => 1, 'class' => 'Codeages\\Plumber\\Example\\Example2Worker'],
+        'ExampleForward' => [
+            'worker_num' => 1,
+            'class' => 'Codeages\Plumber\ForwardWorker',
+            'destination' => [
+                'host' => 'localhost', // 转发目的(host, port)队列，不能跟当前worker所监听的队列一样。
+                'port' => 11301,
+                'tubeName' => 'ExampleForward',
+            ]
+        ],
     ],
 
     'log_path' => __DIR__ . '/tmp/plumber.log',